extendSequence() with mixins
